Implement separate test_utils module, increase row count
